Did the makority of the work in making War Drive "opaque" to the SDL headers
The idea being that I don't want Warp Drive AND SDL to be included by
projects using Warp Drive. It's been done for everything, using forward
declarations and moving the SDL includes to the sources. The only issue
remaining is SDL_Events. Since each GameState needs to implement
handling of its own events the solution won't be as simple. A better
solution will have to be thought as I still like the idea of "hiding SDL
away" so to speak

+++ WarpDrive/display/sprite.cpp
@@ -139,11 +139,11 @@ int Sprite::playAnimation(int i)
 	return current.X()+1;
 }
 
-void Sprite::setDrawAlpha(unsigned int alpha)
+void Sprite::setDrawAlpha(uint8_t alpha)
 {
 	for(unsigned int i = 0; i < sheets.size(); i++)
 	{
-        sheets[i]->setDrawAlpha( static_cast<Uint8>(alpha) );
+        sheets[i]->setDrawAlpha( static_cast<uint8_t>(alpha) );
 	}
 }
